+++ src/Kit/Navigation/SideMenu/index.tsx
@@ -1,0 +1,4 @@
+import InternalSideMenu from './SideMenu';
+
+export const KitSideMenu = InternalSideMenu;
+KitSideMenu.displayName = 'KitSideMenu';