+++ src/Kit/Navigation/index.ts
@@ -5,3 +5,4 @@ export {KitHeader} from './Header';
 export {KitMenu} from './Menu';
 export {KitSteps} from './Steps';
 export {KitPagination} from './Pagination';
+export {KitSideMenu} from './SideMenu';